feat: Add Embla Carousel autoplay feature
The code changes include adding the Embla Carousel autoplay feature by installing the "embla-carousel-autoplay" package. This feature allows the carousel to automatically transition between slides after a specified delay.

Recent user commits:
- chore: Update npm dependencies for Clerk and Supabase libraries
- feat: Add sign-up and sign-in pages
- chore: Update prettier configuration with tailwindcss plugin
- feat: update readme (#16)
- feat: update readme (#15)
- feat: Update technologies and contributors in readme.md (#14)
- feat: Remove console.log statement in ChatSection.tsx (#13)
- feat: Add types for AppExplanResponseFromStream and ExplanationResponse (#12)
- Implement loading (#11)
- Merge pull request #10 from buisihung11/fix/terraform-log-stream

Recent repository commits:
- chore: Update npm dependencies for Clerk and Supabase libraries
- feat: Add sign-up and sign-in pages
- chore: Update prettier configuration with tailwindcss plugin
- feat: update readme (#16)
- feat: update readme (#15)
- feat: Update technologies and contributors in readme.md (#14)
- feat: Remove console.log statement in ChatSection.tsx (#13)
- feat: Add types for AppExplanResponseFromStream and ExplanationResponse (#12)
- Implement loading (#11)
- Merge pull request #10 from buisihung11/fix/terraform-log-stream
