+++ frontend/package.json
@@ -76,6 +76,7 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
+    "embla-carousel-autoplay": "^8.1.3",
     "embla-carousel-react": "^8.0.2",
     "framer-motion": "^11.1.7",
     "input-otp": "^1.2.4",
