+++ frontend/src/components/Auth/AuthStyle.tsx
@@ -1,0 +1,599 @@
+const TopRightShine: React.FC<{ className?: string }> = ({ className }) => (
+  <svg
+    className={className}
+    width="529"
+    height="499"
+    viewBox="0 0 529 499"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter2_f_8026_110)">
+      <ellipse
+        cx="524.924"
+        cy="22.6709"
+        rx="22.3794"
+        ry="381.284"
+        fill="url(#paint2_linear_8026_110)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter3_f_8026_110)">
+      <ellipse
+        cx="524.924"
+        cy="-177.946"
+        rx="22.3794"
+        ry="180.667"
+        fill="url(#paint3_linear_8026_110)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter4_f_8026_110)">
+      <ellipse
+        cx="409.694"
+        cy="40.8839"
+        rx="22.25"
+        ry="381.5"
+        transform="rotate(15 409.694 40.8839)"
+        fill="url(#paint4_linear_8026_110)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter5_f_8026_110)">
+      <ellipse
+        cx="467.558"
+        cy="-176.031"
+        rx="321.5"
+        ry="187.5"
+        transform="rotate(15 467.558 -176.031)"
+        fill="url(#paint5_linear_8026_110)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }}>
+      <ellipse
+        cx="491.369"
+        cy="-264.896"
+        rx="160.5"
+        ry="95.5"
+        transform="rotate(15 491.369 -264.896)"
+        fill="url(#paint6_linear_8026_110)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }}>
+      <ellipse
+        cx="489.493"
+        cy="-257.893"
+        rx="135"
+        ry="80.25"
+        transform="rotate(15 489.493 -257.893)"
+        fill="url(#paint7_linear_8026_110)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <defs>
+      <filter
+        id="filter0_f_8026_110"
+        x="97.2672"
+        y="-423.485"
+        width="477.307"
+        height="686.892"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_110"
+        />
+      </filter>
+      <filter
+        id="filter1_f_8026_110"
+        x="274.448"
+        y="-445.321"
+        width="338.241"
+        height="744.685"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_110"
+        />
+      </filter>
+      <filter
+        id="filter2_f_8026_110"
+        x="413.545"
+        y="-447.613"
+        width="222.759"
+        height="940.567"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_110"
+        />
+      </filter>
+      <filter
+        id="filter3_f_8026_110"
+        x="413.545"
+        y="-447.613"
+        width="222.759"
+        height="539.335"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_110"
+        />
+      </filter>
+      <filter
+        id="filter4_f_8026_110"
+        x="219.622"
+        y="-416.663"
+        width="380.145"
+        height="915.093"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_110"
+        />
+      </filter>
+      <filter
+        id="filter5_f_8026_110"
+        x="3.19824"
+        y="-525.391"
+        width="928.719"
+        height="698.72"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="75"
+          result="effect1_foregroundBlur_8026_110"
+        />
+      </filter>
+      <linearGradient
+        id="paint0_linear_8026_110"
+        x1="335.921"
+        y1="-373.384"
+        x2="335.921"
+        y2="213.307"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint1_linear_8026_110"
+        x1="443.569"
+        y1="-366.229"
+        x2="443.569"
+        y2="220.271"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint2_linear_8026_110"
+        x1="524.924"
+        y1="-358.613"
+        x2="524.924"
+        y2="403.954"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint3_linear_8026_110"
+        x1="524.924"
+        y1="-358.613"
+        x2="524.924"
+        y2="2.72144"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint4_linear_8026_110"
+        x1="409.694"
+        y1="-340.616"
+        x2="409.694"
+        y2="422.384"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint5_linear_8026_110"
+        x1="467.558"
+        y1="-363.531"
+        x2="467.558"
+        y2="11.4689"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint6_linear_8026_110"
+        x1="491.369"
+        y1="-360.396"
+        x2="491.369"
+        y2="-169.396"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint7_linear_8026_110"
+        x1="489.493"
+        y1="-338.143"
+        x2="489.493"
+        y2="-177.643"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+    </defs>
+  </svg>
+)
+
+const TopLeftShine: React.FC<{ className?: string }> = ({ className }) => (
+  <svg
+    className={className}
+    width="659"
+    height="537"
+    viewBox="0 0 659 537"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <g style={{ mixBlendMode: "color-dodge" }} filter="url(#filter1_f_8026_43)">
+      <ellipse
+        cx="262.378"
+        cy="-50.076"
+        rx="26.5"
+        ry="293.25"
+        transform="rotate(5 262.378 -50.076)"
+        fill="url(#paint1_linear_8026_43)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter2_f_8026_43)">
+      <ellipse
+        cx="359.106"
+        cy="29.9931"
+        rx="22.3794"
+        ry="381.284"
+        transform="rotate(-10 359.106 29.9931)"
+        fill="url(#paint2_linear_8026_43)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter3_f_8026_43)">
+      <ellipse
+        cx="324.269"
+        cy="-167.576"
+        rx="22.3794"
+        ry="180.667"
+        transform="rotate(-10 324.269 -167.576)"
+        fill="url(#paint3_linear_8026_43)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter4_f_8026_43)">
+      <ellipse
+        cx="168.788"
+        cy="67.9388"
+        rx="22.25"
+        ry="381.5"
+        transform="rotate(5 168.788 67.9388)"
+        fill="url(#paint4_linear_8026_43)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }} filter="url(#filter5_f_8026_43)">
+      <ellipse
+        cx="188.107"
+        cy="-155.729"
+        rx="321.5"
+        ry="187.5"
+        transform="rotate(5 188.107 -155.729)"
+        fill="url(#paint5_linear_8026_43)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }}>
+      <ellipse
+        cx="196.124"
+        cy="-247.379"
+        rx="160.5"
+        ry="95.5"
+        transform="rotate(5 196.124 -247.379)"
+        fill="url(#paint6_linear_8026_43)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <g style={{ mixBlendMode: "lighten" }}>
+      <ellipse
+        cx="195.494"
+        cy="-240.156"
+        rx="135"
+        ry="80.25"
+        transform="rotate(5 195.494 -240.156)"
+        fill="url(#paint7_linear_8026_43)"
+        fill-opacity="0.5"
+      />
+    </g>
+    <defs>
+      <filter
+        id="filter0_f_8026_43"
+        x="-39.0848"
+        y="-403.13"
+        width="388.448"
+        height="729.588"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_43"
+        />
+      </filter>
+      <filter
+        id="filter1_f_8026_43"
+        x="136.633"
+        y="-431.219"
+        width="251.489"
+        height="762.287"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_43"
+        />
+      </filter>
+      <filter
+        id="filter2_f_8026_43"
+        x="200.306"
+        y="-434.518"
+        width="317.6"
+        height="929.023"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_43"
+        />
+      </filter>
+      <filter
+        id="filter3_f_8026_43"
+        x="196.926"
+        y="-434.542"
+        width="254.687"
+        height="533.932"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_43"
+        />
+      </filter>
+      <filter
+        id="filter4_f_8026_43"
+        x="39.8235"
+        y="-401.115"
+        width="257.93"
+        height="938.107"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="44.5"
+          result="effect1_foregroundBlur_8026_43"
+        />
+      </filter>
+      <filter
+        id="filter5_f_8026_43"
+        x="-282.594"
+        y="-494.631"
+        width="941.402"
+        height="677.805"
+        filterUnits="userSpaceOnUse"
+        color-interpolation-filters="sRGB"
+      >
+        <feFlood flood-opacity="0" result="BackgroundImageFix" />
+        <feBlend
+          mode="normal"
+          in="SourceGraphic"
+          in2="BackgroundImageFix"
+          result="shape"
+        />
+        <feGaussianBlur
+          stdDeviation="75"
+          result="effect1_foregroundBlur_8026_43"
+        />
+      </filter>
+      <linearGradient
+        id="paint0_linear_8026_43"
+        x1="155.139"
+        y1="-331.681"
+        x2="155.139"
+        y2="255.01"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint1_linear_8026_43"
+        x1="262.378"
+        y1="-343.326"
+        x2="262.378"
+        y2="243.174"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint2_linear_8026_43"
+        x1="359.106"
+        y1="-351.29"
+        x2="359.106"
+        y2="411.277"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint3_linear_8026_43"
+        x1="324.269"
+        y1="-348.243"
+        x2="324.269"
+        y2="13.0914"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint4_linear_8026_43"
+        x1="168.788"
+        y1="-313.561"
+        x2="168.788"
+        y2="449.439"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint5_linear_8026_43"
+        x1="188.107"
+        y1="-343.229"
+        x2="188.107"
+        y2="31.7714"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint6_linear_8026_43"
+        x1="196.124"
+        y1="-342.879"
+        x2="196.124"
+        y2="-151.879"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint7_linear_8026_43"
+        x1="195.494"
+        y1="-320.406"
+        x2="195.494"
+        y2="-159.906"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+    </defs>
+  </svg>
+)
+
+
+export { TopRightShine, TopLeftShine }
