+++ frontend/src/components/Auth/Logo.tsx
@@ -1,0 +1,7 @@
+const Logo = () => (
+  <p className="text-pretty bg-gradient-to-r from-white via-white to-white/30 bg-clip-text text-3xl leading-10 text-transparent">
+    SyarGPT
+  </p>
+)
+
+export default Logo
