+++ frontend/src/components/Auth/index.ts
@@ -1,0 +1,5 @@
+import { TopLeftShine, TopRightShine } from "./AuthStyle"
+import Logo from "./Logo"
+import FeatureCarousel from "./FeatureCarousel"
+
+export { TopLeftShine, TopRightShine, Logo, FeatureCarousel }
